Fix npm test
