+++ package.json
@@ -24,9 +24,9 @@
   },
   "dependencies": {
     "chalk": "2.3.2",
-    "cross-spawn": "^5.1.0",
+    "cross-spawn": "5.1.0",
     "generator-jhipster": ">=4.13.3",
-    "jsonfile": "^4.0.0",
+    "jsonfile": "4.0.0",
     "mkdirp": "0.5.1",
     "modify-package-dependencies": "^1.0.0",
     "semver": "5.5.0",
